Allow webhook CA bundle to be taken from secret (#273)
Previously, if you are providing your own certs for the webhook, the CA bundle
had to be specified as a helm chart parameter. This made it harder to deploy
the operator if the cert was generated through a helm chart too because you
won't know what the cert is when you deploy your app.

This change is to allow the CA cert to be read from the webhook.tlsSecret. If
that secret has the key ca.crt, then it will be patched in the webhook config
by the operator.

The secret that contains the tlsSecret has to be provided to the operator now.
An empty string implies the operator will generate the certificate internally.

Support for deploying without the webhook had to be fixed as it was broken in
the prior change that relaxed the cert-manager dependency.

The default name of the tls secret when using cert-manager was updated to
reflect the name chosen when installing with olm.

Some e2e cleanup was added. The custom-cert-webhook had some logging testing
that was moved to its own test. Then I created a copy of custom-cert-webhook
for when the CA bundle is in the secret.