Allow catalog path to be specified
This adds to the VerticaDB API the ability to specify the catalog path.
The new parameter name is: spec.local.catalogPath.

Previously, the catalog directory was the same as the data directory.
The rationale for splitting this out now is to make it easier to migrate
existing databases to k8s. If the database was created outside of k8s
and a custom catalog path was specified then it wasn't possible to run
that database inside k8s.

For backwards compatability, if the catalog path isn't specified,
then we assume the catalog path is the same as the data path.

Due to environment variables changing in the subcluster statefulsets,
upgrading to this version of the operator will cause a restart of each
of the Vertica pods.