Allow local paths to share the same mount point
This is a bug fix for when the local paths (data, depot, catalog) all
share the same mount location. Previous to this change, the pods in a
statefulset would fail to start. It failed to start because they all
inadvertently shared the same mount. This fixes this by adding a mount
only if the local paths differ.