Allow local paths to share the same mount point (#275)
This is a bug fix for when the local paths (data, depot, catalog) all share the
same mount location. Previous to this change, the pods in a statefulset will
fail to start if the paths are all the same. It fails to start because they all
shared the same mount. This fixes it by adding a mount only if the local paths
differ.