Operator pod readiness probe to wait for webhook (#280)
The controller-runtime added a health check that will return only when the
webhook is up. However, it isn't enabled by default. Instead we just use the
normal 'ping' checking, which returns as soon as the pod is running. This
commit is to change the readiness probe to use the one in controller-runtime
that waits for the webhook.

As part of this, I removed the manual webhook check that we use for the e2e
tests.