Add the webhook wait
I thought we had resolved the issue the health check and the webhook.
But we are getting some instability in the e2e tests because we are
applying vdb before the webhook is ready. The new health check we use
closes the gap but it doesn't solve the timing issue. It is still
possible for the operator pod to be ready before the webhook is ready.
This adds back the explicit wait.