Separate PVC cleanup from VerticaDB cleanup
During an e2e run, I have seen the test harness (kuttl) hang when it
does testcase cleanup. We typically delete the PVC and VerticaDB at the
same time. Due to timing, it is possible for the PVC to be recreated;
this can happen if the sts controller recreated pods that were just
deleted. This causes kuttl to hang because the PVC that it tried to
delete is not gone due to the recreate. The solution is to split
deletion of the VerticaDB and the PV into separate test steps. The
VerticaDB is deleted first, wait for it to finish, then delete the PVC.