Installer to create http conf dir if enabled
This fixes a bug that was introduced earlier in this release when we did
the podfacts refactor. We were failing to create the http conf dir in
the installer reconciler if the http server is enabled. The plan is
enable the http e2e tests, which would have caught this, when we use the
12.0.2 server release in the CI.