Allow authorization to /metrics endpoint with TLS certificates (#287)
The RBAC auth proxy is a sidecar that runs alongside the operator manager to
control access to the operators /metrics endpoint. Previous to this change, if
you used the auth proxy, the only authentication method was with bearer tokens
that you can get from the service account. This change allows custom client
certificates to be used for authentication also.

To get this behaviour, the certificates need to be added to a secret, then
referenced in the new helm chart parameter: prometheus.tlsSecret. The secret
must have the following keys set: tls.key, tls.crt, and ca.crt.

By using custom certificates, authentication to the rbac proxy can be done with
a curl command like this:

curl --key tls.key --cert tls.crt --cacert ca.crt https://verticadb-operator-metrics-service.<namespace>.svc:8443/metrics

If the prometheus.tlsSecret is omitted, which is the default behaviour, you can
only use the bearer token to authenticate with the RBAC auth proxy.

I also include a change in here to make it easier to setup the RBAC rules for
the metrics endpoint. The prometheus.createRBACProxy parameter created the RBAC
rules. We had a rule that you could not use this parameter when deploying
several instances of the operator. This has been relaxed now. As a result, the
default value for prometheus.createRBACProxy is now true.

This also adds a deprecation warning for the prometheus.createServiceMonitor
parameter. This will be removed in a future release.