vdb-gen to handle db that have no communal auth
There are a couple of cases where a Vertica database can be created
without storing authentication parms for communal:
- using HDFS as the communal store
- using AWS S3, but using IAM profiles for authentication

vdb-gen wasn't handling this case well. For AWS S3, it was expecting to
see the authentication parms. It failed when it couldn't find it. And
it handles the HDFS case, but still prints out an empty credential
secret when it dumps the manifests.

This fixes both scenarios in vdb-gen.