Expose the http port in the service object (#290)
The http server is experimental at the moment and is hidden behind
spec.httpServerMode: Enabled

This exposes a new port (8443) in the service port for the http server that
runs inside vertica server. This also removes the agent port in the service
object, which was never used because we never ended up running the agent within
the container.

I am also renaming the spec.httpServerSecret field in the VerticaDB to
spec.httpServerTLSSecret. Since this feature is experimental, there should not
be anyone using it, so no dependency on the field that I am renaming.