Fix timing in k-safety-0-scaling e2e test (#293)
Small timing window when in k-safety-0-scaling. We check that we have
rebalanced after adding new nodes. But adding nodes and rebalance are handled
by separate reconciler in the operator. So it's possible that we have completed
the add node, but haven't run the rebalance, and the e2e test checks the state
and fails. Solution is to add a step before the rebalance check that makes sure
the operator has completed its reconcile iteration. This is a hard one to see
in the diff because I had to renumber some of the test steps too.