Cleanup subcluster metrics on VerticaDB delete (#299)
We have a callback that cleans up Prometheus metrics the operator maintains
whenever the VerticaDB is deleted. Previously, we weren't able to remove any
metrics that had the subcluster name as a label. This fixes that. We upgrade
the prometheus client to take advantage of a new API DeletePartialMatch, that
will allow us to delete those metrics.