Include db name in Prometheus metrics
Adding a new label to all of the Prometheus metrics that the operator
collects. We include the database name: e.g. db_name="vertdb". I am
adding this to be in-sync with the server as it uses the same label.
This will help allow us to pull metrics into the same dashboard for
operator and server.