Don't start metric ep if metrics are disabled
When the metrics are disabled (through helm chart parmeter
`prometheus.expose=Disable`), we weren't closing off the metrics
endpoint. It was still being started on port :8080. There is an option
to the operator called `--metrics-bind-address`. We were omitting this
when metrics were disabled, but we ended up picking the default for this
option and starting the metrics serving anyway. This fixes it by
changing the default value for that parameter to "0", which prevents the
metrics server from starting.