Change the db label to revive_instance_id
This changes earlier commits in the PR. I know use the
revive_instance_id as a unique identifier of metrics for a cluster. This
will be in-sync with the server metrics, so we will be able to associate
operator metrics with the server with the same life of the cluster.