Include subcluster_oid in metrics (#304)
This changes a few prometheus metrics to include the subcluster oid instead of
the name. The subcluster name can change where as the oid stays constant. So,
it's better to build a dashboard around. This is to sync with what the server
is doing, so we will be able to combine metrics from the operator and the
server in a single dashboard.

In order to include the subcluster oid, we needed to collect it. The podfacts
were updated and a new status field was added to store the oid. The status
field is necessary so that we know the oid if vertica is down and we can't
query anything.