Allow workflow dispach to set the name of the e2e run (#305)
This takes advantage of a new feature in GitHub workflows that allows you to
set the name of the run. We'll be able to use this to distinguish the
dispatched runs from the Vertica internal CI: is it a release we are testing,
is it the daily server, developer branch, etc.