Split e2e tests into separate legs (#307)
We are going to split the tests/e2e and tests/e2e-extra directory into more
directories. This will allow us to reduce the runtime of the CI run. Some of
the same tests run with different communal back ends, but I don't see much
value in this -- we haven't caught any problems that are specific to a communal
back end.