Allow posix path for communal storage (#308)
This relaxes some checks in the webhook so that we can use a POSIX path for
communal storage. This should be used only for development purposes. A
supported communal path (e.g. S3, HDFS, etc.) should still be used if deploying
a production environment. This is to aid in development as it will allow for
faster deployments as we don’t need a communal storage service running such as
minio. Instead, if you are running on a single host you can use a host path to
share communal storage across pods.

Also, I am adding a back door to allow the operator to create and manage an
enterprise database. You can get into this mode if you set the shard count in
the CR to zero. There is a webhook rule that prevents this, so you will need to
disable the webhook to run it this way. This is being added for internal dev
testing and is by no way an officially supported use of the operator.