Allow the readinessProbe to be configured (#309)
The VerticaDB currently has a hard coded timeout for the readiness probe of 1
second. It does a 'select 1' query to vertica to check that its up and ready to
accept connections. If this does not respond in 1 second, and this happens 3
times in a row, then we assume the endpoint isn't up. Routing rules from the
service object will no longer reach that pod until the readiness probe is
successful again.

A new knob is added to the VerticaDB to allow the readiness probe to be
altered. See the readinessProbeOverride parameter in the example below. The new
parameter is the corev1.Probe type, so it can accept any values that are
typically included with probes.

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: v
spec:
...
  readinessProbeOverride:
    timeoutSeconds: 5
    periodSeconds: 20