Rename some fields added in current release
Some new fields in the VerticaDB were added this release for the Vertica
http server. There is going to be two http servers in the future, so
changing the names of the fields to accomadate for this future change.
This rename is fine because we haven't yet release a version of the
operator that had the old names. They are new in this release.