Use container for helm ut (#312)
Occasionally, we see failures when trying to install the unittest plugin for
helm. Helm pulls it from github, but occasionally it can fail with errors such
as this:

helm plugin install https://github.com/quintush/helm-unittest
Support linux-amd64
Retrieving https://api.github.com/repos/quintush/helm-unittest/releases/tags/v0.2.11
No download_url found only searching for linux
Downloading  to location /tmp/_dist/
curl: Remote file name has no length!
curl: (23) Failed writing received data to disk/application

Ideally, we would like to retry because this seems to be an intermittent
problem. But the curl command is embedded in helm. So, instead I'm switching to
use a container as that seems a bit more robust. As a side benefit, this also
allows us to fix the version of the plugin we are using.