Add an init program to vertica-k8s image (#320)
This adds a proper init program to the vertica-k8s container so that it
can reap zombie process and automatically restart cron/sshd if they
fail. We use [s6-overlay](https://github.com/just-containers/s6-overlay)
as our init program. It adds about 7MB (uncompressed) to our images.

There were a few additional items included in this due to that change:
- Removed the /config directory setup that was in the container entry
script. This is better suited for the operator when it runs the install
reconciler. We already did some setup in the operator, this just moves
all of the pieces that were previously in the container.
- Related to the above, we set the securityPolicy.fsGroup so that the PV
that is mounted in the container uses the gid of dbadmin. This gives it
write access to the PV and no chmod was needed for the PVs in the
container.
- Cleanup of the pod_facts.go. We kept track of certain files and
directories with individual bools. This is managed in a much better way
through a map.
- Removed some hacks that were added for 11.0.0 related to directory
permissions of /opt/vertica/config/logrotate. The new minimum version
the operator supports is now 11.0.1.