Add startupProbe and livenessProbe for vertica server (#325)
This adds new health probes to better handle when vertica crashes or
becomes unresponsive. This was handled before internally by the
operator. It would kill vertica process that were no longer responsive.
This approach wasn't configurable and didn't handle some edge cases
well. Using health probes is better because it's a more idiomatic way of
doing things in k8s. They can also be tuned with new parameters in the
VerticaDB.

The following probes are added:
- startupProbe: We wait for up to 20-minutes for the vertica process to
start. The probe will monitor the startup.log. This can be tuned using
the `spec.startupProbeOverride` field in the VerticaDB.
- livenessProbe: This probe will wait for about 1.5-minutes before
failing. The probe checks if the TCP client port is open. This differs
from the readinessProbe, as that checks the 'select 1' canary query. We
chose a different path to minimize variability as this probe isn't
forgiving if it fails (we reschedule the pod). This can be tuned using
the `spec.livenessProbeOverride` field in the VerticaDB

The operator will only ever restart vertica when the startupProbe is
active. The livenessProbe is too short. If we started vertica when that
probe is active, we risk having the pod get rescheduled during the
restart.

When moving up to an operator with this fix, the operator will add the
new probes. This requires restart of the pods. The operator will handle
this automatically, but you will see a rolling update as the change is
applied.