Allow revive when local paths aren't known (#332)
This will make it easier to initialize a VerticaDB with revive. It takes
the guesswork out of the local paths. It will do a pre-pass for the
revive to get the local paths that existed for the database. If the
paths differ from what is in the VerticaDB, the operator will update the
paths to match.

The webhook had to be relaxed to allow changing of the paths. Before,
this was strictly prevented. Now, we allow the local paths to change as
long as the database hasn't been initialized yet.

There are no new externals for this change. The revive pre-pass is
always done. If the paths are already correct, there is no change in
behavior.