Use a deterministic pod for AT commands
This ensures we try to use the same pod when calling various AT
commands. When picking a pod, we always sort all pods that qualify and
then use the one with the lowest statefulset index.

This is necessary for remove node. If we end up using a pod that is
actually being removed, we run the risk the livenessProbe could actually
delete the pod that is running the AT commands.  This bug isn't exposed
in older release of the operator because the livenessProbe is new to
this current release.