Revive to handle DB names that have capitals (#334)
This fixes a regression in the new revive handler work. It wasn't
handling database names that had capital letters in them. When we run
the revive with --display-only option, we would erronously think the
paths are bad because we weren't checking for the correct case.

Updates an e2e test to use capital letters in its db name that would
have caught this sooner.

This regression exists only in the current release. So, no changie entry
exists.