Allow revive if some data paths differ (#336)
This PR will allow a revive if the data paths are different among the
vertica hosts. Vertica databases managed in k8s assume the data paths
are the same across all pods. This relaxes that to allow a revive if the
data path is either equal to `spec.local.dataPath` or
`spec.local.catalogPath`. Both paths are still backed by the PV, so will
persist between restarts.

In order to allow the data paths to differ, the db existence check in
pod facts now checks the catalog path.

Note: the data path could mismatch due to a bug in revive. When revive
reconstructs the admintools.conf, it will set the data path to be the
same as the catalog path. This happens even if they originally were
different. In that case, when the database does a scaling operation, the
new vertica hosts will have catalog and data to be the same. But
existing vertica hosts will have catalog and data that will differ. I
included an e2e test to mimic this behaviour.

Closes #335