Fix case sensitivity of repositories (#340)
This fixes a problem when running the CI in repositories that have uppercase
letters in the GitHub name. The CI will push images to ghcr.io. It uses the
github ID as part of the name component (e.g. ghcr.io/<github id>/<image>:<tag>).
However, docker doesn't allow uppercase letters in the name component. This
fixes that by using lowercase names for the name component.