Default httpServerMode will setup http server
This changes the default value for spec.httpServerMode. If it's an empty
string, or auto, we will setup the http server if running with a vertica
version of 12.0.3 or higher.

This did some refactor of the version package. To allow the VerticaDB to
check if version is compatible, some of the logic of the version package
was moved into the VerticaDB API.