Create custom image that does not have ssh keys stored (#343)
This removes most keys from all images.

The operator requires the dbadmin and root account to have valid keys. This is
needed to support a cluster with mixed releases as admintools uses ssh to
communicate between hosts. So, they must be preserved and kept static.

We offer a new version of the image with the dbadmin and root account keys
gone. This can be built with a new variable (NO_SSH_KEYS). When using this
version of the image, you must add your own ssh keys into a secret and provide
it in the VerticaDB spec.sshSecret field.

The GitHub action (e2e-leg-1) have been modified to use that custom image.