save space in image by removing duplicate files
When a file is changed in a subsequent layer, the previous version still
ends up getting download.  Thus, an upgrade operation will result in
duplicate files in different layers.  The fix for this is to squash all
the layers after the upgrade which will eliminate all the duplicate
files.

The downside to that is that everything goes into one image and the
download cannot be done in parallel.  So, I moved the update+squash to
the beginning so that new layers from COPY and apt-get can be downloaded
in parallel.