Fix bug preventing restart when pods have sidecars (#345)
This fixes a regression that was introduced in 1.10.0 of the operator.
If you ran your VertiacDB with sidecars, it is possible that the
operator fails to restart vertica. It will be stuck in a loop waiting
for the startupProbe to expire.

The operator was checking the pod's status.containerStatuses list to see
if the startupProbe is active. It assumed (wrongly) that it only ever
needed to look at the first entry in that list. So, it kept looking at
the wrong entry and can make the wrong inference about whether the
startupProbe is still active.

A new e2e test was added to mimic the error and verify the fix.