Always gen http cert if httpServerMode is auto
The http cert is needed if running httpServerMode. When this is set to
Auto, we use to rely on the vertica version to know if it supports the
http server.

That was causing some issues in our e2e tests. A lot of reconcile
iterations were needed for created db. Eventually some of the tests were
going into a few minute lag between reconcile iterations due to the
exponential backoff.

This change is to create the cert ahead of any version check. This may
mean we generate the cert and put it in the sts more often. But this
http server will required at some point in the future, so this shouldn't
be too much of a problem.