Remove HDFS testing from CI
We were relying on an external image (uhopper/hadoop-datanode:2.8.1),
which has since been removed. I could have tried to regenerated this
image myself. However, the HDFS testing has been somewhat flaky. And I
don't see a great value add in putting the effort in to add it back.

The e2e tests that use to run on hdfs where changed to use hostpath. The
name of the test in the CI was renamed to e2e-leg-2.

Note, the operator still supports HDFS as the communal storage. This
commit is strictly about removing the CI testing for HDFS.