Setup keys for client side agent access (#353)
In order to use the agent within the operator, we require certain keys
to be setup in well known locations under /opt/vetrica. This will setup
those keys to allow for client side agent access.

A change to the vertica image is necessary as we had to save off the
keys that were generated in install_vertica. This applies to all but the
nosshkeys image. This follows the same pattern we do for licensing and
logrotate state.