UP node check needs to be made more resilient
This fixes an issue with subcluster creation. We weren't running the
rebalance_shards() to add subscriptions to the new subcluster. To hit
this you had to have a few specific things:
- had to use a v11 server
- the database has to be migrated from enterprise to eon

There was a server issue during migration, to be fixed separately, that
set some wrong state for certain tables. Some tables were identified as
being "shared", which to the server means we need active shard
subscriptions to query them. If no subscriptions, then the query would
fail with "ERROR 9099:  Cannot find participating nodes to run the
query".

This was affecting queries the operator does to catalog tables -- key
for this fix was it affected any query to the nodes table. Because of
this, the operator deemed the new scaled out nodes as down. So, it would
never run the rebalance.

The fix is to align the UP node check with the livenessProbe by looking
to see if the vertica process is running. We still query out the node
state from node, but it is used for information purposes now.