Support subcluster names with underscore (#362)
Before this, a subcluster's name couldn't contain the underscore
character because it is part of sts and pod names and according to FQDN
rules, an underscore is not a valid character.

This adds better support by simply mapping '_' to a '-' before building
pod/service/sts name. This will allow us to support the default
subcluster name in vertica – default_subcluster.

The webhook now allows subcluster name with '_'. However, two
subclusters could have have identical sts names: e.g. default_subcluster
and default-subcluster. We reject that case now in the webhook.

---------

Co-authored-by: Matt Spilchen <matt.spilchen@vertica.com>