Fix intermittent failure in schedule-only e2e test
The schedule-only test can fail occastionally. It's a combination of
running add node outside of the operator and the operator attempting to
restart the very same node we are adding. I'm avoiding this timing
scenario in the test by setting autoRestartVertica to false until we
specifically want to test restart (at step 60).