Fix helm install without cluster admin priv
Normally, when installing the operator via a helm chart you need full
cluster admin privileges. There is a path that you can take install the
operator if you don't have cluster admin priv. That was broken that this
PR is fixing.

The root cause was that cluster roles/rolebindings were added to the
operator so that it could patch the webhook config objects. This is only
needed when the operator generates its own cert for the webhook. We are
adding conditional logic in those manifests to only apply them if the
operator is generating the certs.

New e2e testcase was added to verify install without cluster admin priv.