Fix operator deployment issues
I missed a few things in my first commit. We patch the webhook CA bundle
for other use cases. Not just when we have the operator generate the
self-signed cert. So we need to broaden when the clusterroles and
clusterrolebindings are added.

The rules to add clusterroles and clusterrolebindings are:
- webhook must be enabled
- webhook cert source is internal
- webhook cert source is a user provided secret