Fix helm install without cluster admin priv (#369)
Normally, when installing the operator via a helm chart you need full
cluster admin privileges. There is a path that you can take install the
operator if you don't have cluster admin priv. That was broken a few
months ago. This PR is fixing that.

The root cause was that cluster roles/rolebindings were added to the
operator so that it could patch the webhook config objects. This is only
needed when cert for the webhook is either given as a secret or the
operator will generate its own. I am adding conditional logic in those
manifests to only apply the cluster role/rolebindings in those cases.

New e2e testcase was added to verify install without cluster admin priv.