Fix udx-cpp e2e test
A change went into the nightly vertica server that modified the sdk
examples for AggregateFunctions.sql. Ideally, we should have different
expected output for the appropriate vertica version. But the nightly
build still reports the vertica version as 12.0.4 -- even though it has
shipped. So, unfortunately all we can do is skip comparing of the
expected output if 12.0.4 or above. We need to come back and revisit
this once the nightly build correctly points to the next version.