Don't strip vertica-k8s image except for minimal (#387)
We call strip against /opt/vertica/bin/vertica to reduce the image size,
but this impacts debugging of the server. For instance, collecting
vstack output will show missing function names if the vertica binary has
been stripped. This changes the image so that only the `-minimal` image
will strip the vertica binary. This adds about 50MB uncompressed / 8MB
compressed to the image size.