Fixes for server version change (#389)
This is a fix for the e2e tests related to the nightly vertica server
test. The latest server builds updated it's version to 23.3.0. This
broke a few things:
- all of the upgrade tests were from 11.1.1. As per the upgrade path,
you cannot go from 11.1.1 to 23.3.0 as you are skipping the 12.0.x
release. The fix here is to change all of the upgrade tests to start
from 12.0.2.
- some of the udx tests have different expected results based on the
server version. The server version check wasn't quite right, so we were
picking the wrong expected result for 23.3.0.