Additional ports added to headless service
This adds new ports to the headless service. This was done for
compatibility with istio. When istio is enabled with TPROXY redirection
and strict mTLS, the ports are needed to ensure vertica pods can
continue to communicate with each other. See this for more info:
https://istio.io/latest/docs/ops/configuration/traffic-management/traffic-routing/#headless-services

The headless service previously only had the ssh port. That was renamed
from ssh to tcp-ssh to follow the istio port naming guidelines (see
https://istio.io/latest/docs/ops/configuration/traffic-management/protocol-selection/)