Start http server if httpServerMode changes in the VerticaDB (#390)
The operator now starts the http server in the vertica pods. When a CR
is created with httpServerMode Disabled, the http server will not be
running. If the CR is changed so that httpServerMode is Enabled, the
operator will start the http service. The only allowed transitions are
from Disabled to Enabled/Auto and Auto to Enabled. Any other will be
rejected.

A new e2e test has been added and tests a transition from `Disabled` to
`Enabled`.

---------

Co-authored-by: Matt Spilchen <matt.spilchen@vertica.com>