Move vertica-k8s to Ubuntu 23.04 (Lunar Lobster) (#393)
An image change for vertica-k8s that addresses some security
vulnerabilities that we found internally:

- change the OS of the final stage of the image to Ubuntu 23.04 (Lunar
Lobster). This newer version gets security updates quicker than the
previous version we were using. Moving to this version addressed a few
vulnerabilities that we had
- remove pip from the image. We had an earlier attempt at doing this,
but there was remnents still around.
- change the builder os version, an early stage in the container, from
centos:centos7.9.2009 to quay.io/centos/centos:stream8. The old image we
were using was deprecated. If you try building the vertica container on
a new mac, it would complain. Switching to a new centos.