Reuse node names when installing new k8s pods (#398)
This corrects node name reuse when the operator installs new hosts in
the admintools.conf. The old reuse logic was a bit too simple and could
end up skip reusing a node name under certain circumstances.

Note, there are two kinds of node names in vertica. The node name before
a node is added to the cluster -- in the format of node####. And the
vnode name, which is the name after its added to the cluster -- in the
format of v_<dbname>_node####. This impacts only the first format.