Auto delete PVC on VerticaDB deletion
When the VerticaDB is deleted, we will also delete all of the PVCs too.
We get this behaviour automatically by filling in the ownerReferences in
the PVC template.

Only new StatefulSet's created by the operator will get this behaviour. If
upgrading the operator to this version, and the StatefulSet isn't recreated,
then auto delete won't happen. We did this because you cannot change the
PVC template in the StatefulSet after it is created.

Signed-off-by: Matt Spilchen <matt.spilchen@vertica.com>