Avoid infinite reconcile if subclusters share svc
This fixes a bug when you have two subclusters share the same service
object. Subclusters can share a service object if they both have the
same value for spec.subclusters[].serviceName. The service object has a
label for the subcluster name (vertica.com/subcluster-name). When more
than one subcluster shared the service object, we would continually
update the service object by cycling through the subcluster names. This
resulted in an infinite reconcile. We would change the service object,
that would trigger another reconcile, that would cause another change
in the service object, ...

A secondary bug fix in this relates to handle of annotations. When we
compare expected vs current service object, we don't handle annotations
correctly. The annotations can be an empty map or nil map. Technically
they are both equivalent, but our comparison flags them as different.
Thus we always try to update the service object. The k8s client is smart
enough not to change the service object in this case. So, we just end up
doing a no-op update. But we can save some time if we treated them the same.

Signed-off-by: Matt Spilchen <matt.spilchen@vertica.com>